requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==2.6.1
influxdb-client==1.44.0
python-dotenv==1.0.0
gunicorn==21.2.0

